fix(calendar): fixing on select time slots & preview
