+++ packages-web/calendar/src/components/CalendarContainer.ts
@@ -1,11 +1,12 @@
-import { Component, createElement, ReactChild, ReactNode } from "react";
 import { hot } from "react-hot-loader/root";
+import { Component, createElement, ReactChild, ReactNode } from "react";
 
 import { Calendar, CalendarEvent } from "./Calendar";
 import { fetchData } from "../utils/data";
 import { Container } from "../utils/namespaces";
 import dateMath from "date-arithmetic";
 import moment from "moment";
+import { validateProps } from "../utils/validation";
 
 export interface CalendarContainerState {
     alertMessage: ReactChild;
@@ -49,9 +50,7 @@ class CalendarContainer extends Component<Container.CalendarContainerProps, Cale
     render(): ReactNode {
         const readOnly = this.isReadOnly();
         const alertMessage =
-            this.state.alertMessage ||
-            CalendarContainer.validateProps(this.props) ||
-            CalendarContainer.validateCustomFormats(this.props);
+            this.state.alertMessage || validateProps(this.props) || CalendarContainer.validateCustomFormats(this.props);
 
         return createElement(
             "div",
@@ -355,11 +354,18 @@ class CalendarContainer extends Component<Container.CalendarContainerProps, Cale
     };
 
     private onClickSlot = (slotInfo: Container.EventInfo) => {
+        const { start, end, slots } = slotInfo;
         mx.data.create({
             entity: this.props.eventEntity,
             callback: newEvent => {
-                newEvent.set(this.props.startAttribute, dateMath.startOf(slotInfo.start, "day"));
-                newEvent.set(this.props.endAttribute, dateMath.endOf(slotInfo.end, "day"));
+                const daysInBetween = dateMath.diff(start, end, "day", false);
+                if (daysInBetween === 0 && slots.length > 1) {
+                    newEvent.set(this.props.startAttribute, start);
+                    newEvent.set(this.props.endAttribute, end);
+                } else {
+                    newEvent.set(this.props.startAttribute, dateMath.startOf(start, "day"));
+                    newEvent.set(this.props.endAttribute, dateMath.endOf(end, "day"));
+                }
                 if (
                     this.props.mxObject &&
                     this.props.newEventContextPath &&
@@ -437,50 +443,6 @@ class CalendarContainer extends Component<Container.CalendarContainerProps, Cale
                 error: error => window.mx.ui.error(`An error occurred while executing the nanoflow: ${error.message}`)
             });
         }
-    }
-
-    static validateProps(props: Container.CalendarContainerProps): ReactChild {
-        const errorMessages: string[] = [];
-
-        if (props.onClickEvent === "callMicroflow" && !props.onClickMicroflow) {
-            errorMessages.push("On click event is set to 'Call a microflow' but no microflow is selected");
-        } else if (props.onClickEvent === "callNanoflow" && !props.onClickNanoflow.nanoflow) {
-            errorMessages.push("On click event is set to 'Call a nanoflow' but no nanoflow is selected");
-        }
-        if (props.onCreate === "callMicroflow" && !props.onCreateMicroflow) {
-            errorMessages.push("On create event is set to 'Call a microflow' but no microflow is selected");
-        } else if (props.onCreate === "callNanoflow" && !props.onCreateNanoflow.nanoflow) {
-            errorMessages.push("On create event is set to 'Call a nanoflow' but no nanoflow is selected");
-        }
-        if (props.onChangeEvent === "callMicroflow" && !props.onChangeMicroflow) {
-            errorMessages.push("On change event is set to 'Call a microflow' but no microflow is selected");
-        } else if (props.onChangeEvent === "callNanoflow" && !props.onChangeNanoflow.nanoflow) {
-            errorMessages.push("On change event is set to 'Call a nanoflow' but no nanoflow is selected");
-        }
-        if (props.dataSource === "microflow" && !props.dataSourceMicroflow) {
-            errorMessages.push("Datasource is set to 'microflow' but no microflow is selected");
-        } else if (props.dataSource === "nanoflow" && !props.dataSourceNanoflow.nanoflow) {
-            errorMessages.push("Datasource is set to 'nanoflow' but no nanoflow is selected");
-        }
-        if (props.dataSource === "context" && props.mxObject && props.mxObject.getEntity() !== props.eventEntity) {
-            errorMessages.push(`${props.friendlyId}: Context entity does not match the event entity`);
-        }
-        if (props.view === "custom" && props.customViews.length <= 0) {
-            errorMessages.push(`${props.friendlyId}: View is set to "custom" but there is no view selected`);
-        }
-        if (props.view === "standard" && (props.defaultView === "work_week" || props.defaultView === "agenda")) {
-            errorMessages.push(`${props.friendlyId}: ${props.defaultView} is only available in custom view`);
-        }
-        if (errorMessages.length) {
-            return createElement(
-                "div",
-                {},
-                "Error in calendar configuration:",
-                errorMessages.map((message, key) => createElement("p", { key }, message))
-            );
-        }
-
-        return "";
     }
 
     static validateCustomFormats(props: Container.CalendarContainerProps): ReactChild {
